refactor : 엔티티 및 VO 패키지 정리
