+++ core/data/src/commonMain/kotlin/com/whatever/caramel/core/data/mapper/AuthMapper.kt
@@ -1,27 +1,20 @@
 package com.whatever.caramel.core.data.mapper
 
-import com.whatever.caramel.core.domain.entity.auth.SocialLoginType
-import com.whatever.caramel.core.domain.entity.user.UserBasic
-import com.whatever.caramel.core.domain.model.aggregate.UserAuthAggregation
-import com.whatever.caramel.core.domain.model.auth.AuthToken
-import com.whatever.caramel.core.remote.dto.auth.LoginPlatform
+import com.whatever.caramel.core.domain.vo.auth.AuthToken
+import com.whatever.caramel.core.domain.vo.auth.UserAuthAggregation
+import com.whatever.caramel.core.domain.vo.user.UserStatus
 import com.whatever.caramel.core.remote.dto.auth.ServiceToken
 import com.whatever.caramel.core.remote.dto.auth.SignInResponse
+import kotlinx.datetime.Instant
 
-fun SocialLoginType.toLoginPlatform(): LoginPlatform =
-    when (this) {
-        SocialLoginType.APPLE -> LoginPlatform.APPLE
-        SocialLoginType.KAKAO -> LoginPlatform.KAKAO
-    }
-
-
-fun SignInResponse.toUserAuthAggregation() : UserAuthAggregation {
+fun SignInResponse.toUserAuthAggregation(): UserAuthAggregation {
     return UserAuthAggregation(
-        userBasic = UserBasic(
-            id = 0L,
-            userStatus = this.userStatus.toUserStatus(),
-            coupleId = this.coupleId
-        ),
+        coupleId = this.coupleId,
+        nickname = this.nickname,
+        userStatus = UserStatus.valueOf(this.userStatus.name),
+        birthDayMillisecond = this.birthDay?.let {
+            Instant.parse(it).toEpochMilliseconds()
+        },
         authToken = serviceToken.toAuthToken()
     )
 }
@@ -31,7 +24,7 @@ fun ServiceToken.toAuthToken() = AuthToken(
     refreshToken = this.refreshToken
 )
 
-fun Pair<String, String>.toAuthToken() : AuthToken {
+fun Pair<String, String>.toAuthToken(): AuthToken {
     return AuthToken(
         accessToken = this.first,
         refreshToken = this.second
