+++ core/data/src/commonMain/kotlin/com/whatever/caramel/core/data/mapper/CoupleMapper.kt
@@ -1,6 +1,6 @@
 package com.whatever.caramel.core.data.mapper
 
-import com.whatever.caramel.core.domain.model.couple.CoupleInvitationCode
+import com.whatever.caramel.core.domain.vo.couple.CoupleInvitationCode
 import com.whatever.caramel.core.remote.dto.couple.CoupleInvitationCodeResponse
 import kotlinx.datetime.Instant
