+++ core/data/src/commonMain/kotlin/com/whatever/caramel/core/data/repository/CoupleRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.whatever.caramel.core.data.repository
 
 import com.whatever.caramel.core.data.mapper.toCoupleInvitationCode
 import com.whatever.caramel.core.data.util.safeCall
-import com.whatever.caramel.core.domain.model.couple.CoupleInvitationCode
+import com.whatever.caramel.core.domain.vo.couple.CoupleInvitationCode
 import com.whatever.caramel.core.domain.repository.CoupleRepository
 import com.whatever.caramel.core.remote.datasource.RemoteCoupleDataSource
 import com.whatever.caramel.core.remote.dto.couple.CoupleConnectRequest
