+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/entity/User.kt
@@ -1,0 +1,22 @@
+package com.whatever.caramel.core.domain.entity
+
+import com.whatever.caramel.core.domain.exception.CaramelException
+import com.whatever.caramel.core.domain.exception.code.AppExceptionCode
+import com.whatever.caramel.core.domain.vo.user.UserMetaData
+import com.whatever.caramel.core.domain.vo.user.UserProfile
+import com.whatever.caramel.core.domain.vo.user.UserStatus
+
+data class User(
+    private val id : Long? = null,
+    val userStatus: UserStatus = UserStatus.NONE,
+    val userProfile: UserProfile? = null,
+    val userMetaData: UserMetaData? = null,
+) {
+    val hasId get() = id != null
+    val requireId get() = id ?: throw CaramelException(
+        code = AppExceptionCode.NULL_VALUE,
+        message = "사용자 ID가 존재하지 않습니다",
+        debugMessage = null
+    )
+}
+
