--- core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/model/aggregate/UserAuthAggregation.kt
@@ -1,9 +1,0 @@
-package com.whatever.caramel.core.domain.model.aggregate
-
-import com.whatever.caramel.core.domain.entity.user.UserBasic
-import com.whatever.caramel.core.domain.model.auth.AuthToken
-
-data class UserAuthAggregation(
-    val authToken : AuthToken,
-    val userBasic : UserBasic
-)
