+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/repository/AuthRepository.kt
@@ -1,8 +1,8 @@
 package com.whatever.caramel.core.domain.repository
 
-import com.whatever.caramel.core.domain.model.aggregate.UserAuthAggregation
-import com.whatever.caramel.core.domain.model.auth.AuthToken
-import com.whatever.caramel.core.domain.usecase.auth.SocialLoginInputModel
+import com.whatever.caramel.core.domain.vo.auth.UserAuth
+import com.whatever.caramel.core.domain.vo.auth.AuthToken
+import com.whatever.caramel.core.domain.vo.auth.SocialLoginType
 
 interface AuthRepository {
     suspend fun loginWithSocialPlatform(inputModel: SocialLoginInputModel)  : UserAuthAggregation
