+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/usecase/user/RefreshUserSessionUseCase.kt
@@ -1,6 +1,6 @@
 package com.whatever.caramel.core.domain.usecase.user
 
-import com.whatever.caramel.core.domain.entity.user.UserStatus
+import com.whatever.caramel.core.domain.vo.user.UserStatus
 import com.whatever.caramel.core.domain.exception.code.AppExceptionCode
 import com.whatever.caramel.core.domain.exception.CaramelException
 import com.whatever.caramel.core.domain.repository.AuthRepository
