+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/auth/UserAuth.kt
@@ -1,0 +1,11 @@
+package com.whatever.caramel.core.domain.vo.auth
+
+import com.whatever.caramel.core.domain.vo.user.UserStatus
+
+data class UserAuth(
+    val coupleId : Long? = null,
+    val userStatus: UserStatus,
+    val nickname : String? = null,
+    val birthDayMillisecond : Long? = null,
+    val authToken : AuthToken,
+)
