+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/couple/CoupleInvitationCode.kt
@@ -1,4 +1,4 @@
-package com.whatever.caramel.core.domain.model.couple
+package com.whatever.caramel.core.domain.vo.couple
 
 data class CoupleInvitationCode(
     val invitationCode: String,
