+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/user/UserMetaData.kt
@@ -1,0 +1,6 @@
+package com.whatever.caramel.core.domain.vo.user
+
+data class UserMetaData(
+    val agreementServiceTerms: Boolean = false,
+    val agreementPrivacyPolicy: Boolean = false
+)
