+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/user/UserProfile.kt
@@ -1,0 +1,7 @@
+package com.whatever.caramel.core.domain.vo.user
+
+data class UserProfile(
+    val nickName : String,
+    val gender : Gender,
+    val birthdayMillisecond : Long
+)
