+++ core/domain/src/commonMain/kotlin/com/whatever/caramel/core/domain/vo/user/UserStatus.kt
@@ -1,4 +1,4 @@
-package com.whatever.caramel.core.domain.entity.user
+package com.whatever.caramel.core.domain.vo.user
 
 enum class UserStatus {
     NONE,
