+++ feature/login/src/commonMain/kotlin/com/whatever/caramel/feature/login/LoginRoute.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.whatever.caramel.core.domain.entity.auth.SocialLoginType
+import com.whatever.caramel.core.domain.vo.auth.SocialLoginType
 import com.whatever.caramel.feature.login.mvi.LoginIntent
 import com.whatever.caramel.feature.login.mvi.LoginSideEffect
 import com.whatever.caramel.feature.login.social.SocialAuthenticator
